derive: track included files with relative path
To make the compiler understand which files contributed to the generated
code, in order to make it aware when a recompilation is needed, we track
the file paths using `include_bytes!()` calls in the generated code.
This can be a problem for reproducible builds, because the working
directory contributes to the file hash.

This PR makes Askama generate relative paths if possible. This is always
possible on Unix systems, but on Windows hosts absolute paths have to be
used if the code is kept on different drives.